added new routes

+++ InterviewExam.Domain/Interfaces/Repositories/IOrderRepository.cs
@@ -4,7 +4,7 @@ namespace InterviewExam.Domain.Interfaces.Repositories
 {
     public interface IOrderRepository
     {
-        public Task<IEnumerable<Order>> GetAsync(long customerId, DateTime? startDate, DateTime? endDate, CancellationToken cancellationToken);
+        public Task<IEnumerable<Order>> GetAsync(int customerId, DateTime? startDate, DateTime? endDate, CancellationToken cancellationToken);
         public Task<Order> CreateAsync(Order newOrder, CancellationToken cancellationToken);
     }
 }
